+++ config/session.php
@@ -1,0 +1,6 @@
+<?php declare(strict_types=1);
+
+return [
+    'name'          => 'slimapp',
+    'cache_expire'  => 0,
+];