+++ config/settings.php
@@ -7,10 +7,11 @@ date_default_timezone_set('UTC');
 $settings = [];
 
 // Register configurations.
-$settings['path']   = require __DIR__ . '/path.php';
-$settings['error']  = require __DIR__ . '/error.php';
-$settings['logger'] = require __DIR__ . '/logger.php';
-$settings['twig']   = require __DIR__ . '/twig.php';
+$settings['path']       = require __DIR__ . '/path.php';
+$settings['error']      = require __DIR__ . '/error.php';
+$settings['logger']     = require __DIR__ . '/logger.php';
+$settings['twig']       = require __DIR__ . '/twig.php';
+$settings['session']    = require __DIR__ . '/session.php';
 
 // Return array to container.
 return $settings;