+++ src/App/Middleware/SessionMiddleware.php
@@ -1,0 +1,41 @@
+<?php
+declare(strict_types=1);
+
+namespace App\Middleware;
+
+use Psr\Http\Message\ResponseInterface;
+use Psr\Http\Message\ServerRequestInterface;
+use Psr\Http\Server\MiddlewareInterface;
+use Psr\Http\Server\RequestHandlerInterface;
+use Symfony\Component\HttpFoundation\Session\Session;
+
+class SessionMiddleware implements MiddlewareInterface
+{
+    /**
+     * @var \Symfony\Component\HttpFoundation\Session
+     */
+    private $session;
+
+    /**
+     * SessionMiddleware constructor.
+     *
+     * @param \Symfony\Component\HttpFoundation\Session\Session $session
+     */
+    public function __construct(Session $session)
+    {
+        $this->session = $session;
+    }
+
+    /**
+     * Start the session.
+     *
+     * @param \Psr\Http\Message\ServerRequestInterface $request
+     * @param \Psr\Http\Server\RequestHandlerInterface $handler
+     * @return \Psr\Http\Message\ResponseInterface
+     */
+    public function process(ServerRequestInterface $request, RequestHandlerInterface $handler): ResponseInterface
+    {
+        $this->session->start();
+        return $handler->handle($request);
+    }
+}