+++ CHANGELOG.md
@@ -3,3 +3,4 @@
     * Configured Slim 4 base.
     * Implement logger.
     * Add Twig template system.
+    * Session management.